- moved the monitor to a more fitting package - removed filestream, no real use for it
